# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.CreateFreelancerProfileReq:
    properties:
      availability:
        type: string
      bio:
        type: string
      contact_email:
        type: string
      currency:
        type: string
      experience:
        type: string
      github_profile:
        type: string
      hourly_rate:
        type: number
      is_available:
        type: boolean
      languages:
        items:
          type: string
        type: array
      linkedin_profile:
        type: string
      portfolio:
        items:
          type: string
        type: array
      preferred_projects:
        items:
          type: string
        type: array
      response_time:
        type: string
      skills:
        items:
          type: string
        type: array
      time_zone:
        type: string
      website:
        type: string
    type: object
  schema.CreateJobApplicationReq:
    properties:
      cover_letter:
        type: string
      currency:
        type: string
      job_id:
        type: string
      message:
        type: string
      proposed_rate:
        type: number
    required:
    - job_id
    type: object
  schema.CreateJobPostingReq:
    properties:
      budget:
        type: number
      budget_type:
        type: string
      contact_email:
        type: string
      currency:
        type: string
      description:
        type: string
      duration:
        type: string
      experience_level:
        type: string
      expires_at:
        type: integer
      location:
        type: string
      skills:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - description
    - title
    type: object
...
  schema.FreelancerProfileResp:
    properties:
      availability:
        type: string
      bio:
        type: string
      bio_html:
        type: string
      client_satisfaction:
        type: number
      completed_projects:
        type: integer
      contact_email:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      experience:
        type: string
      github_profile:
        type: string
      hourly_rate:
        type: number
      id:
        type: string
      is_available:
        type: boolean
      is_verified:
        type: boolean
      languages:
        items:
          type: string
        type: array
      linkedin_profile:
        type: string
      portfolio:
        items:
          type: string
        type: array
      preferred_projects:
        items:
          type: string
        type: array
      response_time:
        type: string
      skills:
        items:
          type: string
        type: array
      time_zone:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
      verification_date:
        type: integer
      website:
        type: string
    type: object
...
  schema.GetFreelancerProfilesResp:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/schema.FreelancerProfileResp'
        type: array
    type: object
  schema.GetJobPostingsResp:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/schema.JobPostingResp'
        type: array
    type: object
...
  schema.HireFreelancerReq:
    properties:
      freelancer_user_id:
        type: string
      message:
        type: string
      subject:
        type: string
    required:
    - freelancer_user_id
    - message
    - subject
    type: object
  schema.HireFreelancerResp:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  schema.JobPostingResp:
    properties:
      application_count:
        type: integer
      budget:
        type: number
      budget_type:
        type: string
      contact_email:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      description:
        type: string
      description_html:
        type: string
      duration:
        type: string
      experience_level:
        type: string
      expires_at:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      skills:
        items:
          type: string
        type: array
      status:
        type: string
      title:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
      views_count:
        type: integer
    type: object
...
  schema.UpdateFreelancerProfileReq:
    properties:
      availability:
        type: string
      bio:
        type: string
      contact_email:
        type: string
      currency:
        type: string
      experience:
        type: string
      github_profile:
        type: string
      hourly_rate:
        type: number
      id:
        type: string
      is_available:
        type: boolean
      languages:
        items:
          type: string
        type: array
      linkedin_profile:
        type: string
      portfolio:
        items:
          type: string
        type: array
      preferred_projects:
        items:
          type: string
        type: array
      response_time:
        type: string
      skills:
        items:
          type: string
        type: array
      time_zone:
        type: string
      website:
        type: string
    required:
    - id
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/freelancer/hire:
    post:
      consumes:
      - application/json
      description: Send hiring message to freelancer
      parameters:
      - description: HireFreelancer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.HireFreelancerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.HireFreelancerResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Hire freelancer
      tags:
      - Freelancer
  /answer/api/v1/freelancer/profile:
    get:
      consumes:
      - application/json
      description: Get freelancer profile
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.FreelancerProfileResp'
              type: object
      summary: Get freelancer profile
      tags:
      - Freelancer
    post:
      consumes:
      - application/json
      description: Create freelancer profile
      parameters:
      - description: CreateFreelancerProfile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.CreateFreelancerProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Create freelancer profile
      tags:
      - Freelancer
    put:
      consumes:
      - application/json
      description: Update freelancer profile
      parameters:
      - description: UpdateFreelancerProfile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateFreelancerProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Update freelancer profile
      tags:
      - Freelancer
  /answer/api/v1/freelancer/profiles:
    get:
      consumes:
      - application/json
      description: Get freelancer profiles
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      - description: skills
        in: query
        name: skills
        type: string
      - description: location
        in: query
        name: location
        type: string
      - description: min_rate
        in: query
        name: min_rate
        type: number
      - description: max_rate
        in: query
        name: max_rate
        type: number
      - description: currency
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetFreelancerProfilesResp'
              type: object
      summary: Get freelancer profiles
      tags:
      - Freelancer
  /answer/api/v1/job/application:
    post:
      consumes:
      - application/json
      description: Create job application
      parameters:
      - description: CreateJobApplication
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.CreateJobApplicationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Create job application
      tags:
      - Job
  /answer/api/v1/job/posting:
    post:
      consumes:
      - application/json
      description: Create job posting
      parameters:
      - description: CreateJobPosting
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.CreateJobPostingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Create job posting
      tags:
      - Job
  /answer/api/v1/job/posting/{id}:
    get:
      consumes:
      - application/json
      description: Get job posting by ID
      parameters:
      - description: job_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.JobPostingResp'
              type: object
      summary: Get job posting
      tags:
      - Job
  /answer/api/v1/job/postings:
    get:
      consumes:
      - application/json
      description: Get job postings
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      - description: skills
        in: query
        name: skills
        type: string
      - description: location
        in: query
        name: location
        type: string
      - description: min_budget
        in: query
        name: min_budget
        type: number
      - description: max_budget
        in: query
        name: max_budget
        type: number
      - description: currency
        in: query
        name: currency
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetJobPostingsResp'
              type: object
      summary: Get job postings
      tags:
      - Job
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"